+++ CAWL/CAWLCombFilter.hpp
@@ -15,7 +15,7 @@
 class CAWLCombFilter : public CAWLSoundModule
 {
 public:
-    virtual void setDelay(unsigned int delay);
+    virtual void setDelay(float delay);
     virtual void setFeedForwardGain(float newGain);
     virtual void setFeedbackGain(float newGain);
     virtual void setMixLevel(float newMixLevel);
